requirements.txt
pymongo
python-dotenv
gremlinpython==3.4.13
loguru